VRT: fix warnings introduced by previous commit
git-svn-id: https://svn.osgeo.org/gdal/trunk@32758 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/vrt/vrtdataset.cpp
@@ -906,7 +906,7 @@ CPLErr VRTDataset::AddBand( GDALDataType eType, char **papszOptions )
 /* -------------------------------------------------------------------- */
         const char* pszImageOffset = CSLFetchNameValueDef(papszOptions, "ImageOffset", "0");
         vsi_l_offset nImageOffset = CPLScanUIntBig(
-                                    pszImageOffset, strlen(pszImageOffset));
+                        pszImageOffset, static_cast<int>(strlen(pszImageOffset)));
 
         int nPixelOffset = nWordDataSize;
         if( CSLFetchNameValue(papszOptions, "PixelOffset") != NULL )
